add log archiving to calamares
cala upstream lacks the ability to archive its log, we also want additional
data stored on the final system (to store more context info, inspired by
ubiquity).
to that end we now have a new helper which creates /var/log/installer/
on the target system and populates with some easy to obtain log data.

- debug is the cala session.log
- media-info is the metadata info of the ISO (i.e. type and version)
- casper.log is /var/log/casper.log (i.e. log of casper for iso spin up)
- syslog is /var/log/syslog of the iso
- initial-status.gz is the compressed original dpkg status of the installed
  system (NOT the ISO!)

all of this is chmodded 600 to avoid accidental sensitive data leakage.
some files are opting for world readable as they have no way of containing
anything sensitive.

currently we do not include an oem id file as cala has no support for it.
we also have no partman, because there is no dedicated partman log.

this doesn't work currently because of
https://github.com/calamares/calamares/issues/954
(is simply noop because we copy everything to the live session's /var/log
 which is in RAM)