actually rename the files back upon removal of calmares-settings
currently we leave lingering file artifacts after the installation

doesn't really cause too big problems, except not being able to re-install
the package on account of the diversion file already existing.
mostly makes openqa fail when trying to install all packages in a repo.

(this is not getting fixed on existing installation as the risk of breaking
something is much higher than the use from removing the left over file)