import confs for all modules we use
cala master changed to not install the default configs anymore. that does
however mean that we need to provide defaults for everything since the
code itself isn't self sufficient as far as general purpose defaults go
that are largely the same for all distros. I am not entirely happy with
how cala behaves right now, but it's better than the implicit default
inheritance we had before.

to prevent errors in the modules, import the stock configs for modules
where we previously relied on the upstream defaults being installed.
these are verbatim with the exception of umount which doesn't install
anything for us, we have our own log archiver specifically for neon